Import of DEXTER/Starlight-0.0200 from CPAN.
gitpan-cpan-distribution: Starlight
gitpan-cpan-version:      0.0200
gitpan-cpan-path:         DEXTER/Starlight-0.0200.tar.gz
gitpan-cpan-author:       DEXTER
gitpan-cpan-maturity:     released

+++ lib/Plack/Handler/Starlight.pm
@@ -3,22 +3,25 @@ package Plack::Handler::Starlight;
 use strict;
 use warnings;
 
-our $VERSION = '0.0100';
+our $VERSION = '0.0200';
 
 use base qw(Starlight::Server);
 
-use POSIX qw(:sys_wait_h);
+use Carp ();
+use Config ();
+use Fcntl ();
+use File::Spec;
+use POSIX ();
 use Plack::Util;
 
-use constant CYGWIN_KILL_PROCESS => $^O eq 'cygwin' && eval { require Win32::Process; 1; };
+use constant HAS_WIN32_PROCESS => $^O eq 'cygwin' && eval { require Win32::Process; 1; };
 
 use constant DEBUG => $ENV{PERL_STARLIGHT_DEBUG};
 
 sub new {
-    my ($klass, %args) = @_;
+    my ($class, %args) = @_;
 
     # setup before instantiation
-    my $listen_sock;
     my $max_workers = 10;
     for (qw(max_workers workers)) {
         $max_workers = delete $args{$_}
@@ -26,7 +29,7 @@ sub new {
     }
 
     # instantiate and set the variables
-    my $self = $klass->SUPER::new(%args);
+    my $self = $class->SUPER::new(%args);
     if ($^O eq 'MSWin32') {
         # forks are emulated
         $self->{is_multithread}  = Plack::Util::TRUE;
@@ -37,18 +40,20 @@ sub new {
         $self->{is_multithread}  = Plack::Util::FALSE;
         $self->{is_multiprocess} = Plack::Util::TRUE;
     };
-    $self->{listen_sock} = $listen_sock
-        if $listen_sock;
     $self->{max_workers} = $max_workers;
 
     $self->{main_process} = $$;
     $self->{processes} = +{};
+
+    $self->{_kill_stalled_processes_delay} = 10;
 
     $self;
 }
 
 sub run {
     my($self, $app) = @_;
+
+    $self->_daemonize();
 
     warn "*** starting main process $$" if DEBUG;
     $self->setup_listener();
@@ -59,7 +64,7 @@ sub run {
         my ($sig) = @_;
         warn "*** SIG$sig received in process $$" if DEBUG;
         local ($!, $?);
-        my $pid = waitpid(-1, WNOHANG);
+        my $pid = waitpid(-1, &POSIX::WNOHANG);
         return if $pid == -1;
         delete $self->{processes}->{$pid};
     };
@@ -73,10 +78,14 @@ sub run {
             warn "*** SIG$sig received in process $$" if DEBUG;
             $self->{term_received}++;
         };
-        while (not $self->{term_received}) {
+        for (my $loop = 0; not $self->{term_received}; $loop++) {
             warn "*** running ", scalar keys %{$self->{processes}}, " processes" if DEBUG;
-            foreach my $pid (keys %{$self->{processes}}) {
-                delete $self->{processes}->{$pid} if not kill 0, $pid;
+            if ($loop >= $self->{_kill_stalled_processes_delay} / ($self->{main_process_delay}||1)) {
+                $loop = 0;
+                # check stalled processes once per n sec
+                foreach my $pid (keys %{$self->{processes}}) {
+                    delete $self->{processes}->{$pid} if not kill 0, $pid;
+                }
             }
             foreach my $n (1 + scalar keys %{$self->{processes}} .. $self->{max_workers}) {
                 $self->_create_process($app);
@@ -91,7 +100,7 @@ sub run {
                 warn "*** stopping process $pid" if DEBUG;
                 kill $sigterm, $pid;
             }
-            if (CYGWIN_KILL_PROCESS) {
+            if (HAS_WIN32_PROCESS) {
                 $self->_sleep(1);
                 foreach my $pid (keys %{$self->{processes}}) {
                     my $winpid = Cygwin::pid_to_winpid($pid) or next;
@@ -104,6 +113,9 @@ sub run {
                 warn "*** waiting for process ", $pid if DEBUG;
                 waitpid $pid, 0;
             }
+        }
+        if ($^O eq 'cygwin' and not HAS_WIN32_PROCESS) {
+            warn "Win32::Process is not installed. Some processes might be still active.\n";
         }
         warn "*** stopping main process $$" if DEBUG;
         exit 0;
@@ -119,6 +131,62 @@ sub run {
             $self->_sleep($self->{spawn_interval});
         }
     }
+}
+
+sub _daemonize {
+    my $self = shift;
+
+    if ($^O eq 'MSWin32') {
+        foreach my $arg (qw(daemonize pid)) {
+            die "$arg parameter is not supported on this platform ($^O)\n" if $self->{$arg};
+        }
+    }
+
+    my ($pidfh, $pidfile);
+    if ($self->{pid}) {
+        $pidfile = File::Spec->rel2abs($self->{pid});
+        if (defined *Fcntl::O_EXCL{CODE}) {
+            sysopen $pidfh, $pidfile, Fcntl::O_WRONLY|Fcntl::O_CREAT|Fcntl::O_EXCL
+                                               or die "Cannot open pid file: $self->{pid}: $!\n";
+        } else {
+            open $pidfh, '>', $pidfile         or die "Cannot open pid file: $self->{pid}: $!\n";
+        }
+    }
+
+    if (defined $self->{error_log}) {
+        open STDERR, '>>', $self->{error_log}  or die "Cannot open error log file: $self->{error_log}: $!\n";
+    }
+
+    if ($self->{daemonize}) {
+
+        chdir File::Spec->rootdir              or die "Cannot chdir to root directory: $!\n";
+
+        open STDIN,  '<', File::Spec->devnull  or die "Cannot open null device for reading: $!\n";
+        open STDOUT, '>', File::Spec->devnull  or die "Cannot open null device for writing: $!\n";
+
+        defined(my $pid = fork)                or die "Cannot fork: $!\n";
+        if ($self->{pid} and $pid) {
+            print $pidfh "$pid\n"              or die "Cannot write pidfile $self->{pid}: $!\n";
+            close $pidfh;
+            exit;
+        }
+
+        close $pidfh if $pidfh;
+
+        if ($Config::Config{d_setsid}) {
+            POSIX::setsid()                    or die "Cannot setsid: $!\n";
+        }
+
+        if (not defined $self->{error_log}) {
+            open STDERR, '>&', \*STDOUT        or die "Cannot dup null device for writing: $!\n";
+        }
+    }
+
+    if ($pidfile) {
+        $self->_add_to_unlink($pidfile);
+    }
+
+    return;
 }
 
 sub _sleep {
